// packages/app/src/account/interface/debt-account-progress-summary.interface.ts
export interface DebtAccountProgressSummaryInterface {
    readonly closedAmount: number;
    readonly creditAmount: number;
    readonly debitAmount: number;
    readonly openedAmount: number;
    readonly outstandingAmount: number;
    readonly paidAmount: number;
    readonly percentage: number;
    readonly totalAmount: number;
}